+++ crates/sevenmark_html/src/render/brace/blockquote.rs
@@ -19,10 +19,9 @@ pub fn render(
 
     let style = utils::build_style(parameters);
     let merged_class = utils::merge_class(classes::BLOCKQUOTE, parameters);
-    let (dk, dark_tag) = utils::dark_style_parts(utils::build_dark_style(parameters));
+    let dk = ctx.add_dark_style(utils::build_dark_style(parameters));
 
     html! {
-        (dark_tag)
         blockquote
             class=(merged_class)
             data-start=[ctx.span_start(span)]