+++ crates/sevenmark_html/src/render/brace/code.rs
@@ -7,13 +7,12 @@ use crate::classes;
 use crate::context::RenderContext;
 use crate::render::utils;
 
-pub fn render(span: &Span, parameters: &Parameters, value: &str, ctx: &RenderContext) -> Markup {
+pub fn render(span: &Span, parameters: &Parameters, value: &str, ctx: &mut RenderContext) -> Markup {
     let lang = utils::get_param(parameters, "lang");
     let style = utils::build_style(parameters);
     let merged_class = utils::merge_class(classes::CODE, parameters);
-    let (dk, dark_tag) = utils::dark_style_parts(utils::build_dark_style(parameters));
+    let dk = ctx.add_dark_style(utils::build_dark_style(parameters));
     html! {
-        (dark_tag)
         pre
             class=(merged_class)
             data-start=[ctx.span_start(span)]