+++ crates/sevenmark_html/src/render/macro/toc.rs
@@ -135,10 +135,9 @@ fn render_styled_label(
 ) -> Markup {
     let merged_class = utils::merge_class(base_class, parameters);
     let style = utils::build_style(parameters);
-    let (dk, dark_tag) = utils::dark_style_parts(utils::build_dark_style(parameters));
+    let dk = ctx.add_dark_style(utils::build_dark_style(parameters));
 
     html! {
-        (dark_tag)
         span class=(merged_class) style=[style] data-dk=[dk] {
             (render_toc_label(children, ctx))
         }
@@ -152,11 +151,10 @@ fn render_ruby_label(
 ) -> Markup {
     let class = utils::merge_class(classes::RUBY, parameters);
     let style = utils::build_style(parameters);
-    let (dk, dark_tag) = utils::dark_style_parts(utils::build_dark_style(parameters));
+    let dk = ctx.add_dark_style(utils::build_dark_style(parameters));
     let ruby_text = utils::get_param(parameters, "ruby").unwrap_or_default();
 
     html! {
-        (dark_tag)
         ruby class=(class) style=[style] data-dk=[dk] {
             (render_toc_label(children, ctx))
             rt { (ruby_text) }